feat(health): enhance health check endpoint with comprehensive monitoring and alerting
- Migrate health check endpoint from Next.js route to Cloudflare Worker function
- Expand health status interface to include degraded state alongside healthy/unhealthy
- Add new service checks: Sentry integration, Worker status, and Performance metrics
- Implement parallel health check execution for improved response times
- Add detailed health status responses with response time and uptime tracking
- Include environment and version information in health check results
- Add critical alert handling for system failures and health check errors
- Implement individual check functions for database, IPFS, RPC, Sentry, Worker, and Performance
- Add comprehensive logging with request metadata and failed service tracking
- Return 200 status for degraded services while maintaining 503 for unhealthy state
- Add response time headers and improved error tracking to health responses
- Consolidate health check logic with better separation of concerns and error handling
