+++ functions/api/health.ts
@@ -1,7 +1,7 @@
 /**
  * Cloudflare Worker: Health Check Endpoint
- * Returns application health status for monitoring
- * GET /api/health - Returns 200 when healthy
+ * Provides comprehensive health status and triggers critical alerts for system failures
+ * GET /api/health - Returns comprehensive health status
  */
 
 import { withErrorTracking } from '../utils/errorTracking';
@@ -15,18 +15,42 @@ interface Env {
 	NEXT_PUBLIC_BASE_RPC_URL?: string;
 	NEXT_PUBLIC_POLYGON_RPC_URL?: string;
 	NEXT_PUBLIC_SKALE_RPC_URL?: string;
+	NEXT_PUBLIC_IPFS_GATEWAY?: string;
+	NEXT_PUBLIC_APP_VERSION?: string;
+}
+
+interface HealthCheckResult {
+	status: 'healthy' | 'degraded' | 'unhealthy';
+	timestamp: string;
+	version: string;
+	environment: string;
+	checks: {
+		database: HealthStatus;
+		ipfs: HealthStatus;
+		rpc: HealthStatus;
+		sentry: HealthStatus;
+		worker: HealthStatus;
+		performance: HealthStatus;
+	};
+	uptime: number;
+	responseTime: number;
 }
 
 interface HealthStatus {
-	status: 'healthy' | 'unhealthy';
-	timestamp: string;
-	version?: string;
-	checks: {
-		database: 'healthy' | 'unhealthy' | 'unknown';
-		ipfs: 'healthy' | 'unhealthy' | 'unknown';
-		rpc: 'healthy' | 'unhealthy' | 'unknown';
-	};
+	status: 'healthy' | 'degraded' | 'unhealthy';
+	message: string;
+	responseTime?: number;
+	lastChecked: string;
+	details?: Record<string, any>;
 }
+
+interface RPCEndpoint {
+	name: string;
+	url: string;
+	chainId: number;
+}
+
+const startTime = Date.now();
 
 const handler: PagesFunction<Env> = async (context) => {
 	const { request, env } = context;
@@ -53,30 +77,55 @@ const handler: PagesFunction<Env> = async (context) => {
 		});
 	}
 
-	const startTime = Date.now();
+	const requestStartTime = Date.now();
 
 	try {
-		// Perform health checks
-		const checks = await performHealthChecks(env);
+		console.log('Health check requested', {
+			userAgent: request.headers.get('user-agent'),
+			cf: (request as any).cf,
+		});
 
-		// Determine overall health status
-		const isHealthy = Object.values(checks).every(
-			(status) => status === 'healthy' || status === 'unknown',
-		);
+		// Run all health checks in parallel
+		const [database, ipfs, rpc, sentry, worker, performance] = await Promise.all([
+			checkDatabase(env),
+			checkIPFS(env),
+			checkRPC(env),
+			checkSentry(env),
+			checkWorker(env),
+			checkPerformance(env, requestStartTime)
+		]);
 
-		const healthStatus: HealthStatus = {
-			status: isHealthy ? 'healthy' : 'unhealthy',
+		const responseTime = Date.now() - requestStartTime;
+		const uptime = Date.now() - startTime;
+
+		const healthResult: HealthCheckResult = {
+			status: determineOverallStatus({ database, ipfs, rpc, sentry, worker, performance }),
 			timestamp: new Date().toISOString(),
-			version: process.env.npm_package_version || '1.0.0',
-			checks,
+			version: env.NEXT_PUBLIC_APP_VERSION || '1.0.0',
+			environment: env.ENVIRONMENT || 'production',
+			checks: { database, ipfs, rpc, sentry, worker, performance },
+			uptime,
+			responseTime
 		};
 
-		const responseTime = Date.now() - startTime;
+		// Handle critical health issues
+		await handleCriticalHealthIssues(healthResult);
 
-		// Return 503 if unhealthy, 200 if healthy
-		const statusCode = isHealthy ? 200 : 503;
+		// Log health check result
+		console.log('Health check completed', {
+			status: healthResult.status,
+			responseTime,
+			uptime,
+			failedChecks: Object.entries(healthResult.checks)
+				.filter(([_, check]) => check.status !== 'healthy')
+				.map(([service, _]) => service)
+		});
 
-		return new Response(JSON.stringify(healthStatus), {
+		// Return appropriate HTTP status code
+		const statusCode = healthResult.status === 'healthy' ? 200 : 
+						  healthResult.status === 'degraded' ? 200 : 503;
+
+		return new Response(JSON.stringify(healthResult), {
 			status: statusCode,
 			headers: {
 				'Content-Type': 'application/json',
@@ -85,18 +134,30 @@ const handler: PagesFunction<Env> = async (context) => {
 				'X-Response-Time': `${responseTime}ms`,
 			},
 		});
-	} catch (error) {
-		console.error('Health check failed:', error);
 
-		const healthStatus: HealthStatus = {
+	} catch (error) {
+		const responseTime = Date.now() - requestStartTime;
+		
+		console.error('Health check failed', { responseTime }, error);
+
+		// Trigger critical alert for health check failure
+		await handleCriticalError(error as Error, env, { responseTime });
+
+		const healthStatus: HealthCheckResult = {
 			status: 'unhealthy',
 			timestamp: new Date().toISOString(),
-			version: process.env.npm_package_version || '1.0.0',
+			version: env.NEXT_PUBLIC_APP_VERSION || '1.0.0',
+			environment: env.ENVIRONMENT || 'production',
 			checks: {
-				database: 'unhealthy',
-				ipfs: 'unknown',
-				rpc: 'unknown',
+				database: { status: 'unhealthy', message: 'Health check failed', lastChecked: new Date().toISOString() },
+				ipfs: { status: 'unhealthy', message: 'Health check failed', lastChecked: new Date().toISOString() },
+				rpc: { status: 'unhealthy', message: 'Health check failed', lastChecked: new Date().toISOString() },
+				sentry: { status: 'unhealthy', message: 'Health check failed', lastChecked: new Date().toISOString() },
+				worker: { status: 'unhealthy', message: 'Health check failed', lastChecked: new Date().toISOString() },
+				performance: { status: 'unhealthy', message: 'Health check failed', lastChecked: new Date().toISOString() }
 			},
+			uptime: Date.now() - startTime,
+			responseTime
 		};
 
 		return new Response(JSON.stringify(healthStatus), {
@@ -111,101 +172,431 @@ const handler: PagesFunction<Env> = async (context) => {
 };
 
 /**
- * Perform health checks for various services
+ * Check database connectivity (KV store)
  */
-async function performHealthChecks(env: Env): Promise<HealthStatus['checks']> {
-	const checks: HealthStatus['checks'] = {
-		database: 'unknown',
-		ipfs: 'unknown',
-		rpc: 'unknown',
-	};
-
-	// Check database connectivity (KV store)
+async function checkDatabase(env: Env): Promise<HealthStatus> {
+	const startTime = Date.now();
+	
 	try {
 		if (env.AUTH_SESSIONS) {
 			// Try to perform a simple KV operation
 			await env.AUTH_SESSIONS.put('health:check', 'ok', { expirationTtl: 60 });
 			await env.AUTH_SESSIONS.get('health:check');
 			await env.AUTH_SESSIONS.delete('health:check');
-			checks.database = 'healthy';
+			
+			const responseTime = Date.now() - startTime;
+			
+			return {
+				status: 'healthy',
+				message: 'Database connection successful',
+				responseTime,
+				lastChecked: new Date().toISOString()
+			};
 		} else {
-			checks.database = 'unknown';
+			return {
+				status: 'degraded',
+				message: 'Database not configured (KV namespace missing)',
+				responseTime: Date.now() - startTime,
+				lastChecked: new Date().toISOString()
+			};
 		}
 	} catch (error) {
-		console.error('Database health check failed:', error);
-		checks.database = 'unhealthy';
+		const responseTime = Date.now() - startTime;
+		
+		return {
+			status: 'unhealthy',
+			message: `Database connection failed: ${(error as Error).message}`,
+			responseTime,
+			lastChecked: new Date().toISOString()
+		};
 	}
+}
 
-	// Check IPFS service status
+/**
+ * Check IPFS connectivity
+ */
+async function checkIPFS(env: Env): Promise<HealthStatus> {
+	const startTime = Date.now();
+	
 	try {
-		// Check if PINATA_JWT is available (indicates IPFS service is configured)
-		const pinataJwt = env.PINATA_JWT || process.env.PINATA_JWT;
+		const ipfsGateway = env.NEXT_PUBLIC_IPFS_GATEWAY || 'https://ipfs.io/ipfs/';
+		const pinataJwt = env.PINATA_JWT;
+		
+		// Test IPFS gateway first
+		const testHash = 'QmUNLLsPACCz1vLxQVkXqqLX5R1X345qqfHbsf67hvA3Nn';
+		const testUrl = `${ipfsGateway}${testHash}`;
+		
+		const controller = new AbortController();
+		const timeoutId = setTimeout(() => controller.abort(), 5000);
+		
+		const gatewayResponse = await fetch(testUrl, {
+			method: 'HEAD',
+			signal: controller.signal
+		});
+		
+		clearTimeout(timeoutId);
+		
+		// Test Pinata API if JWT is available
+		let pinataStatus = 'unknown';
 		if (pinataJwt) {
-			// Perform a simple API call to Pinata to check connectivity
-			const response = await fetch('https://api.pinata.cloud/data/testAuthentication', {
-				method: 'GET',
-				headers: {
-					Authorization: `Bearer ${pinataJwt}`,
-				},
-			});
-
-			if (response.ok) {
-				checks.ipfs = 'healthy';
-			} else {
-				checks.ipfs = 'unhealthy';
+			try {
+				const pinataResponse = await fetch('https://api.pinata.cloud/data/testAuthentication', {
+					method: 'GET',
+					headers: {
+						Authorization: `Bearer ${pinataJwt}`,
+					},
+				});
+				pinataStatus = pinataResponse.ok ? 'healthy' : 'unhealthy';
+			} catch {
+				pinataStatus = 'unhealthy';
 			}
-		} else {
-			checks.ipfs = 'unknown';
 		}
-	} catch (error) {
-		console.error('IPFS health check failed:', error);
-		checks.ipfs = 'unhealthy';
-	}
-
-	// Check RPC endpoint status for supported networks
-	try {
-		const rpcEndpoints = [
-			env.NEXT_PUBLIC_ETHEREUM_RPC_URL || process.env.NEXT_PUBLIC_ETHEREUM_RPC_URL,
-			env.NEXT_PUBLIC_BASE_RPC_URL || process.env.NEXT_PUBLIC_BASE_RPC_URL,
-			env.NEXT_PUBLIC_POLYGON_RPC_URL || process.env.NEXT_PUBLIC_POLYGON_RPC_URL,
-			env.NEXT_PUBLIC_SKALE_RPC_URL || process.env.NEXT_PUBLIC_SKALE_RPC_URL,
-		].filter(Boolean);
-
-		if (rpcEndpoints.length > 0) {
-			// Test at least one RPC endpoint
-			const testEndpoint = rpcEndpoints[0];
-			const response = await fetch(testEndpoint!, {
-				method: 'POST',
-				headers: {
-					'Content-Type': 'application/json',
-				},
-				body: JSON.stringify({
-					jsonrpc: '2.0',
-					method: 'eth_blockNumber',
-					params: [],
-					id: 1,
-				}),
-			});
-
-			if (response.ok) {
-				const data = await response.json();
-				if (data.result) {
-					checks.rpc = 'healthy';
-				} else {
-					checks.rpc = 'unhealthy';
+		
+		const responseTime = Date.now() - startTime;
+		
+		if (gatewayResponse.ok) {
+			return {
+				status: pinataStatus === 'unhealthy' ? 'degraded' : 'healthy',
+				message: `IPFS gateway accessible, Pinata: ${pinataStatus}`,
+				responseTime,
+				lastChecked: new Date().toISOString(),
+				details: { 
+					gateway: ipfsGateway, 
+					pinataStatus,
+					gatewayStatus: gatewayResponse.status
 				}
-			} else {
-				checks.rpc = 'unhealthy';
-			}
+			};
 		} else {
-			checks.rpc = 'unknown';
+			return {
+				status: 'degraded',
+				message: `IPFS gateway returned ${gatewayResponse.status}, Pinata: ${pinataStatus}`,
+				responseTime,
+				lastChecked: new Date().toISOString(),
+				details: { 
+					gateway: ipfsGateway, 
+					pinataStatus,
+					gatewayStatus: gatewayResponse.status
+				}
+			};
 		}
 	} catch (error) {
-		console.error('RPC health check failed:', error);
-		checks.rpc = 'unhealthy';
+		const responseTime = Date.now() - startTime;
+		
+		return {
+			status: 'unhealthy',
+			message: `IPFS check failed: ${(error as Error).message}`,
+			responseTime,
+			lastChecked: new Date().toISOString()
+		};
+	}
+}
+
+/**
+ * Check RPC endpoints
+ */
+async function checkRPC(env: Env): Promise<HealthStatus> {
+	const startTime = Date.now();
+	
+	const rpcEndpoints: RPCEndpoint[] = [
+		{
+			name: 'Ethereum',
+			url: env.NEXT_PUBLIC_ETHEREUM_RPC_URL || 'https://eth.llamarpc.com',
+			chainId: 1
+		},
+		{
+			name: 'Base',
+			url: env.NEXT_PUBLIC_BASE_RPC_URL || 'https://mainnet.base.org',
+			chainId: 8453
+		},
+		{
+			name: 'Polygon',
+			url: env.NEXT_PUBLIC_POLYGON_RPC_URL || 'https://polygon.llamarpc.com',
+			chainId: 137
+		}
+	];
+
+	const results = await Promise.allSettled(
+		rpcEndpoints.map(async (endpoint) => {
+			const controller = new AbortController();
+			const timeoutId = setTimeout(() => controller.abort(), 3000);
+			
+			try {
+				const response = await fetch(endpoint.url, {
+					method: 'POST',
+					headers: { 'Content-Type': 'application/json' },
+					body: JSON.stringify({
+						jsonrpc: '2.0',
+						method: 'eth_chainId',
+						params: [],
+						id: 1
+					}),
+					signal: controller.signal
+				});
+				
+				clearTimeout(timeoutId);
+				
+				if (response.ok) {
+					const data = await response.json();
+					const chainId = parseInt(data.result, 16);
+					
+					return {
+						name: endpoint.name,
+						status: chainId === endpoint.chainId ? 'healthy' : 'degraded',
+						chainId,
+						expectedChainId: endpoint.chainId
+					};
+				} else {
+					return {
+						name: endpoint.name,
+						status: 'unhealthy',
+						error: `HTTP ${response.status}`
+					};
+				}
+			} catch (error) {
+				clearTimeout(timeoutId);
+				return {
+					name: endpoint.name,
+					status: 'unhealthy',
+					error: (error as Error).message
+				};
+			}
+		})
+	);
+
+	const responseTime = Date.now() - startTime;
+	const healthyCount = results.filter(result => 
+		result.status === 'fulfilled' && result.value.status === 'healthy'
+	).length;
+	
+	const totalCount = results.length;
+	const healthyPercentage = (healthyCount / totalCount) * 100;
+
+	let status: 'healthy' | 'degraded' | 'unhealthy';
+	let message: string;
+
+	if (healthyPercentage >= 80) {
+		status = 'healthy';
+		message = `${healthyCount}/${totalCount} RPC endpoints healthy`;
+	} else if (healthyPercentage >= 50) {
+		status = 'degraded';
+		message = `${healthyCount}/${totalCount} RPC endpoints healthy (degraded)`;
+	} else {
+		status = 'unhealthy';
+		message = `${healthyCount}/${totalCount} RPC endpoints healthy (critical)`;
 	}
 
-	return checks;
+	return {
+		status,
+		message,
+		responseTime,
+		lastChecked: new Date().toISOString(),
+		details: {
+			endpoints: results.map((result, index) => ({
+				name: rpcEndpoints[index]?.name || 'Unknown',
+				url: rpcEndpoints[index]?.url || 'Unknown',
+				...(result.status === 'fulfilled' && result.value ? 
+					{ status: result.value.status, chainId: result.value.chainId, expectedChainId: result.value.expectedChainId } : 
+					{ status: 'error', error: result.status === 'rejected' ? result.reason : 'Unknown error' })
+			}))
+		}
+	};
+}
+
+/**
+ * Check Sentry connectivity
+ */
+async function checkSentry(env: Env): Promise<HealthStatus> {
+	const startTime = Date.now();
+	
+	try {
+		const sentryDsn = env.SENTRY_DSN;
+		
+		if (!sentryDsn) {
+			return {
+				status: 'degraded',
+				message: 'Sentry DSN not configured',
+				responseTime: Date.now() - startTime,
+				lastChecked: new Date().toISOString()
+			};
+		}
+
+		// Test Sentry connectivity by parsing DSN
+		try {
+			const url = new URL(sentryDsn);
+			const responseTime = Date.now() - startTime;
+			
+			return {
+				status: 'healthy',
+				message: 'Sentry DSN configured and valid',
+				responseTime,
+				lastChecked: new Date().toISOString(),
+				details: { host: url.host }
+			};
+		} catch {
+			return {
+				status: 'unhealthy',
+				message: 'Sentry DSN invalid format',
+				responseTime: Date.now() - startTime,
+				lastChecked: new Date().toISOString()
+			};
+		}
+	} catch (error) {
+		const responseTime = Date.now() - startTime;
+		
+		return {
+			status: 'unhealthy',
+			message: `Sentry check failed: ${(error as Error).message}`,
+			responseTime,
+			lastChecked: new Date().toISOString()
+		};
+	}
+}
+
+/**
+ * Check Cloudflare Worker status
+ */
+async function checkWorker(env: Env): Promise<HealthStatus> {
+	const startTime = Date.now();
+	
+	try {
+		// Basic worker health checks
+		const workerData = {
+			environment: env.ENVIRONMENT || 'production',
+			timestamp: Date.now(),
+			available: true,
+		};
+
+		const responseTime = Date.now() - startTime;
+
+		return {
+			status: 'healthy',
+			message: `Worker running in ${workerData.environment}`,
+			responseTime,
+			lastChecked: new Date().toISOString(),
+			details: workerData
+		};
+	} catch (error) {
+		const responseTime = Date.now() - startTime;
+		
+		return {
+			status: 'unhealthy',
+			message: `Worker check failed: ${(error as Error).message}`,
+			responseTime,
+			lastChecked: new Date().toISOString()
+		};
+	}
+}
+
+/**
+ * Check performance metrics
+ */
+async function checkPerformance(env: Env, requestStartTime: number): Promise<HealthStatus> {
+	const startTime = Date.now();
+	
+	try {
+		const responseTime = Date.now() - requestStartTime;
+		const checkTime = Date.now() - startTime;
+
+		let status: 'healthy' | 'degraded' | 'unhealthy';
+		let message: string;
+
+		// Check response time thresholds
+		if (responseTime < 100) {
+			status = 'healthy';
+			message = `Performance optimal: ${responseTime}ms response time`;
+		} else if (responseTime < 500) {
+			status = 'degraded';
+			message = `Performance degraded: ${responseTime}ms response time`;
+		} else {
+			status = 'unhealthy';
+			message = `Performance critical: ${responseTime}ms response time`;
+		}
+
+		return {
+			status,
+			message,
+			responseTime: checkTime,
+			lastChecked: new Date().toISOString(),
+			details: {
+				responseTime,
+				checkTime,
+				environment: env.ENVIRONMENT,
+				thresholds: {
+					responseTime: { good: 100, degraded: 500 },
+				},
+			}
+		};
+	} catch (error) {
+		const responseTime = Date.now() - startTime;
+		
+		return {
+			status: 'unhealthy',
+			message: `Performance check failed: ${(error as Error).message}`,
+			responseTime,
+			lastChecked: new Date().toISOString()
+		};
+	}
+}
+
+/**
+ * Determine overall health status
+ */
+function determineOverallStatus(checks: HealthCheckResult['checks']): 'healthy' | 'degraded' | 'unhealthy' {
+	const statuses = Object.values(checks).map(check => check.status);
+	
+	if (statuses.includes('unhealthy')) {
+		return 'unhealthy';
+	} else if (statuses.includes('degraded')) {
+		return 'degraded';
+	} else {
+		return 'healthy';
+	}
+}
+
+/**
+ * Handle critical health issues
+ */
+async function handleCriticalHealthIssues(healthResult: HealthCheckResult): Promise<void> {
+	const criticalChecks = Object.entries(healthResult.checks).filter(
+		([_, check]) => check.status === 'unhealthy'
+	);
+
+	if (criticalChecks.length > 0) {
+		const criticalServices = criticalChecks.map(([service, _]) => service);
+		const error = new Error(`Critical health check failures: ${criticalServices.join(', ')}`);
+		
+		console.error('Critical health check failures detected', {
+			failedServices: criticalServices,
+			healthStatus: healthResult.status,
+			checks: healthResult.checks,
+			uptime: healthResult.uptime
+		}, error);
+	}
+
+	// Also alert on degraded RPC or IPFS (critical for DAO operations)
+	const criticalServices = ['rpc', 'ipfs'];
+	const degradedCriticalServices = criticalServices.filter(
+		service => healthResult.checks[service as keyof typeof healthResult.checks]?.status === 'degraded'
+	);
+
+	if (degradedCriticalServices.length > 0) {
+		console.warn('Critical services degraded', {
+			degradedServices: degradedCriticalServices,
+			healthStatus: healthResult.status
+		});
+	}
+}
+
+/**
+ * Handle critical errors
+ */
+async function handleCriticalError(error: Error, env: Env, context: any): Promise<void> {
+	console.error('Critical error in health check', {
+		error: error.message,
+		stack: error.stack,
+		context,
+		environment: env.ENVIRONMENT
+	});
 }
 
 export const onRequest = withErrorTracking<Env>(handler);
